Fix unnecessary `i32` number size bottleneck
It was not cutting it for normal amounts of total expenditure, and was generally mostly pointless as it ended up cast to `f32` immediately

+++ src/allowances.rs
@@ -15,7 +15,7 @@ pub const DEFAULT_ACCRUAL_DAYS: f32 = 4.0;
 const MILLISECONDS_PER_DAY: u64 = 1000 * 60 * 60 * 24;
 
 pub async fn get_max_allowance_millidollars(daily_allowance: u32, accrual_days: f32) -> f32 {
-	nanodollars_to_millidollars((daily_allowance as f32 * accrual_days) as i32)
+	nanodollars_to_millidollars(daily_allowance as f32 * accrual_days)
 }
 
 async fn time_to_full(executor: &Pool<Sqlite>, user: UserId) -> Option<DateTime<Utc>> {
@@ -125,8 +125,8 @@ const MILLIDOLLARS_PER_NANODOLLAR: f32 = 1.0e6;
 // 	(DEFAULT_DAILY_ALLOWANCE * DEFAULT_ACCRUAL_DAYS) as f32 / MILLIDOLLARS_PER_NANODOLLAR;
 
 /// Converts an integer number of nanodollars to a float number of millidollars, rounded to 2 decimal places.
-pub fn nanodollars_to_millidollars(allowance: i32) -> f32 {
-	let millidollars = allowance as f32 / MILLIDOLLARS_PER_NANODOLLAR;
+pub fn nanodollars_to_millidollars(allowance: f32) -> f32 {
+	let millidollars = allowance / MILLIDOLLARS_PER_NANODOLLAR;
 	(millidollars * PRECISION_MULTIPLIER).round() / PRECISION_MULTIPLIER
 }
 
@@ -139,9 +139,8 @@ pub async fn command_check(
 ) -> Result<(), ()> {
 	let allowance =
 		check_allowance(executor, interaction.user.id, daily_allowance, accrual_days).await;
-	let millidollars = nanodollars_to_millidollars(allowance);
-	let max_millidollars =
-		nanodollars_to_millidollars((daily_allowance as f32 * accrual_days) as i32);
+	let millidollars = nanodollars_to_millidollars(allowance as f32);
+	let max_millidollars = nanodollars_to_millidollars(daily_allowance as f32 * accrual_days);
 	let content = format!(
 		"You have {} out of {} millidollars left.",
 		millidollars, max_millidollars
@@ -155,7 +154,7 @@ pub fn register() -> CreateCommand {
 	CreateCommand::new("allowance").description("Check your current allowance for using ChatGPT.")
 }
 
-async fn get_expenditure(executor: &Pool<Sqlite>, user: Option<UserId>) -> u32 {
+async fn get_expenditure(executor: &Pool<Sqlite>, user: Option<UserId>) -> u64 {
 	if let Some(user) = user {
 		let user_id = user.get() as i64;
 		query!(
@@ -172,7 +171,7 @@ async fn get_expenditure(executor: &Pool<Sqlite>, user: Option<UserId>) -> u32 {
 		.await
 		.unwrap()
 		.cost
-		.map(|n| n as u32)
+		.map(|n| n as u64)
 	} else {
 		query!(
 			"
@@ -186,7 +185,7 @@ async fn get_expenditure(executor: &Pool<Sqlite>, user: Option<UserId>) -> u32 {
 		.await
 		.unwrap()
 		.cost
-		.map(|n| n as u32)
+		.map(|n| n as u64)
 	}
 	.unwrap_or(0)
 }
@@ -203,7 +202,7 @@ pub async fn command_expenditure(
 		.and_then(|option| option.value.as_bool())
 		.unwrap_or(false);
 	let expenditure = get_expenditure(executor, (!all).then_some(interaction.user.id)).await;
-	let millidollars = nanodollars_to_millidollars(expenditure as i32);
+	let millidollars = nanodollars_to_millidollars(expenditure as f32);
 	let content = if !all {
 		format!("You have used {} millidollars.", millidollars)
 	} else {
